gravity and physics work with observers

+++ action.cpp
@@ -4,7 +4,7 @@ Action::Action(actionType_t type)
 : type(type)
 {}
 
-SpeedIncreaseAction::SpeedIncreaseAction(int speed, int dir)
+SpeedIncreaseAction::SpeedIncreaseAction(float speed, int dir)
 : Action (SPEED_INCREASE), speed(speed), dir(dir)
 {}
